requirements.txt
aiohttp>=3.9,<4
aiortc>=1.9,<2
numpy>=1.24
python-dotenv>=1.0
av>=12.0
piper-tts>=1.4
scipy>=1.11
faster-whisper>=1.0
anthropic>=0.40
openai>=1.30
httpx>=0.27
duckduckgo-search>=5.0
google-auth>=2.20
